formating distillery card component
