+++ react-app/src/components/Distilleries/DistilleryCard.js
@@ -1,36 +1,15 @@
-import React, { useEffect } from "react";
-import { useSelector, useDispatch } from "react-redux";
-import { loadDistilleries } from "../../store/distilleries";
+import React from "react";
+
 import styles from "./Distilleries.module.css";
 
-const DistilleryCard = () => {
-    const dispatch = useDispatch();
-
-    const distilleries = useSelector((state) =>
-        Object.values(state.distilleries)
-    );
-    console.log(distilleries, "!!!!!!!!!!!!!!!!!!!!");
-
-    useEffect(() => {
-        dispatch(loadDistilleries());
-    }, [dispatch]);
-
-    if (!distilleries) {
-        return null;
-    }
+const DistilleryCard = ({ distillery }) => {
+    if (!distillery) return null;
     return (
         <div className={styles.distilleryContainer}>
-            {distilleries &&
-                distilleries?.map((distillery) => (
-                    <div>
-                        <div className={styles.distName}>
-                            {distillery?.name}
-                        </div>
-                        <div className={styles.distName}>
-                            {distillery?.street}
-                        </div>
-                    </div>
-                ))}
+            <div className={styles.indivDist}>
+                <div className={styles.distName}>{distillery.name}</div>
+                <div className={styles.distName}>{distillery.street}</div>
+            </div>
         </div>
     );
 };
